requirements.txt
pytorch==2.3.1